+++ server/modules/authenticator/PAM/pam_client_session.cc
@@ -4,7 +4,7 @@
  * Use of this software is governed by the Business Source License included
  * in the LICENSE.TXT file and at www.mariadb.com/bsl11.
  *
- * Change Date: 2025-08-17
+ * Change Date: 2025-09-20
  *
  * On the date above, in accordance with the Business Source License, use
  * of this software will be governed by version 2 or later of the General
@@ -34,7 +34,7 @@ namespace
  * @param output Password output
  * @return True on success, false if packet didn't have a valid header
  */
-bool store_client_password(GWBUF* buffer, mariadb::ClientAuthenticator::ByteVec* output)
+bool store_client_password(GWBUF* buffer, mariadb::AuthByteVec* output)
 {
     bool rval = false;
     uint8_t header[MYSQL_HEADER_LEN];